POLIO-614: update budget status dropdown
