+++ plugins/polio/js/src/constants/budget.ts
@@ -1,0 +1,58 @@
+export const BUDGET_STATES = [
+    'budget_requested',
+    'who_sent_budget',
+    'unicef_sent_budget',
+    'gpei_consolidation',
+    'submitted_to_rrt',
+    'feedback_sent_to_gpei',
+    're_submitted_to_rrt',
+    'submission_to_orpg_operations_1',
+    'feedback_sent_to_rrt1',
+    'submitted_to_orpg',
+    'feedback_sent_to_rrt2',
+    're_submitted_to_orpg',
+    'submission_to_orpg_operations_2',
+    'feedback_sent_to_rrt3',
+    're_submission_to_orpg_operations_2',
+    'submitted_for_approval',
+    'feedback_sent_to_orpg_operations_unicef',
+    'feedback_sent_to_orpg_operations_who',
+    'approved_by_who',
+    'approved_by_unicef',
+    'approved',
+];
+
+export const BUDGET_REQUEST = [
+    'budget_requested',
+    'who_sent_budget',
+    'unicef_sent_budget',
+    'gpei_consolidation',
+];
+
+export const RRT_REVIEW = [
+    'submitted_to_rrt',
+    'feedback_sent_to_gpei',
+    're_submitted_to_rrt',
+    'submission_to_orpg_operations_1',
+];
+
+export const ORPG_REVIEW = [
+    'feedback_sent_to_rrt1',
+    'submitted_to_orpg',
+    'feedback_sent_to_rrt2',
+    're_submitted_to_orpg',
+    'submission_to_orpg_operations_2',
+    'feedback_sent_to_rrt3',
+    're_submission_to_orpg_operations_2',
+];
+
+export const REVIEW_FOR_APPROVAL = [
+    'submitted_for_approval',
+    'feedback_sent_to_orpg_operations_unicef',
+    'feedback_sent_to_orpg_operations_who',
+    'approved_by_who',
+    'approved_by_unicef',
+    'approved',
+];
+
+export const WORKFLOW_SUFFIX = '_at_WFEDITABLE';
