+++ plugins/polio/js/src/styles/constants.js
@@ -21,28 +21,3 @@ export const initialDistrict = {
     opacity: '1',
     zIndex: '1',
 };
-export const BUDGET_STATES = [
-    'budget_requested',
-    'who_sent_budget',
-    'unicef_sent_budget',
-    'gpei_consolidation',
-    'submitted_to_rrt',
-    'feedback_sent_to_gpei',
-    're_submitted_to_rrt',
-    'submission_to_orpg_operations_1',
-    'feedback_sent_to_rrt1',
-    'submitted_to_orpg',
-    'feedback_sent_to_rrt2',
-    're_submitted_to_orpg',
-    'submission_to_orpg_operations_2',
-    'feedback_sent_to_rrt3',
-    're_submission_to_orpg_operations_2',
-    'submitted_for_approval',
-    'feedback_sent_to_orpg_operations_unicef',
-    'feedback_sent_to_orpg_operations_who',
-    'approved_by_who',
-    'approved_by_unicef',
-    'approved',
-];
-
-export const WORKFLOW_SUFFIX = '_at_WFEDITABLE';
